[refactor] 게시글 상세 모달창 - custom hook, import정리
