+++ app/_hooks/useQueries/community/index.ts
@@ -1,6 +1,14 @@
 import { useQuery } from "@tanstack/react-query";
-import { QUERY_KEY_COMMUNITYLIST } from "@/app/_api/queryKeys";
-import { getCommunityList } from "@/app/_api/community/community-api";
+
+import {
+  QUERY_KEY_COMMUNITYLIST,
+  QUERY_KEY_COMMUNITY_POST,
+} from "@/app/_api/queryKeys";
+
+import {
+  getCommunityList,
+  getPostContents,
+} from "@/app/_api/community/community-api";
 
 export const useGetCommunityList = () => {
   const {
@@ -13,4 +21,17 @@ export const useGetCommunityList = () => {
   });
 
   return { communityList, isLoading, isError };
+};
+
+export const useGetPostContents = (post_id: string) => {
+  const {
+    data: communityPost,
+    isLoading: isPostLoading,
+    isError: isPostError,
+  } = useQuery({
+    queryKey: [QUERY_KEY_COMMUNITY_POST, post_id],
+    queryFn: () => getPostContents(post_id),
+  });
+
+  return { communityPost, isPostLoading, isPostError };
 };
